feat(core): move serializer.h
